Split energy ratio result to separate file
